+++ src/components/input/use-input.ts
@@ -33,7 +33,7 @@ export const useInput = (
 
     const inputTextClasses = classNames(
       'spr-block spr-w-full spr-px-size-spacing-2xs spr-py-size-spacing-4xs spr-rounded-border-radius-md spr-outline-none spr-ring-0',
-      'spr-text-color-strong spr-font-size-200',
+      'spr-text-color-strong spr-font-size-200 [font-weight:inherit]',
       'spr-border spr-border-solid',
       'placeholder:spr-text-mushroom-300',
       {